Letterspace Function
Adds a function for use that allows the user to add or remove the space between letters. Note, curve input from user is used for proof of concept. Note, image is resized. Carrying over existing image size should be implemented. imagettfbbox() and possibly preciseMeasure() should be used to accomplish this.