Letterspace Function Integrated
Note, see letterspace_function branch for reference before viewing this.

Letterspace function is no longer dependent on input from the curve
slider. Has it's own slider, and in catalog/products/product
customization admin is able to allow or disallow letterspace, and set a
default letterspace. Letterspace conflicts with curve so logic is put in
to default to curve. Install file is given as reference for integrating
letterspace into sql, but is only two columns and would likely be easier
done by hand through phpmyadmin. Note that line for making compatible
for my version of prestashop is included on
views/templates/hook/DisplayCustomDesignPanel.tpl on line 340. This is
not needed if using a compatible version of PrestaShop.